Fixes a race condition
I ran into a race condition with this. When `currentVisit` gets
re-assigned there are no more references to the JavascriptVisit
it used to hold on to. This will cause it to be deallocated
and the weak ref on the `WebViewBridge` will become null.

The request can finish before the next visit is started and
the new JavascriptVisit is assigned as the delegate.